requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-Login==0.6.2
Flask-WTF==1.1.1
Flask-Cors==4.0.0
psycopg2-binary==2.9.7
SQLAlchemy==2.0.21
WTForms==3.0.1
email-validator==2.0.0
python-dotenv==1.0.0
pandas==2.0.3
openpyxl==3.1.2
Werkzeug==2.3.7
Jinja2==3.1.2
click==8.1.7
python-dateutil==2.8.2
PyJWT==2.8.0
cryptography==41.0.3
reportlab==4.0.7
requests==2.31.0
gunicorn==21.2.0
